Fix undefined behavior.
This will avoid unaligned access if GIPFELI_NO_UNALIGNED is defined,
even on platforms where unaligned access is assumed to be acceptable.
It also fixes a few locations which previously assumed unaligned
access instead of using the UNALIGNED_STORE64 macro, and avoids
right-shifting a 64-bit value by 64 bits, which is undefined behavior.